Chore: Refactor connector code
