+++ packages/connector-chromium/src/connector.ts
@@ -1,13 +1,15 @@
 import compact = require('lodash/compact');
 import * as puppeteer from 'puppeteer-core';
 
-import { contentType, debug as d, dom, HTMLElement, network, HTMLDocument, HttpHeaders } from '@hint/utils';
+import { debug as d, dom, HTMLElement, network, HTMLDocument, HttpHeaders } from '@hint/utils';
 import { normalizeHeaders, Requester } from '@hint/utils-connector-tools';
-import { IConnector, Engine, FetchError, FetchEnd, NetworkData, Request } from 'hint';
+import { IConnector, Engine, NetworkData } from 'hint';
 import { launch, close } from './lib/lifecycle';
 
-const { getContentTypeData, getType } = contentType;
-const { createHTMLDocument, getElementByUrl, traverse } = dom;
+import { getFavicon } from './lib/get-favicon';
+import { onRequestHandler, onRequestFailedHandler, onResponseHandler } from './lib/events';
+
+const { createHTMLDocument, traverse } = dom;
 const { isRegularProtocol } = network;
 const debug: debug.IDebugger = d(__filename);
 
@@ -42,191 +44,10 @@ export default class ChromiumConnector implements IConnector {
             handleSIGHUP: false,
             handleSIGINT: false,
             handleSIGTERM: false,
-            headless: true
+            headless: false
         };
 
         this._options = Object.assign({}, defaults, options);
-    }
-
-    private async createFetchEndPayload(response: puppeteer.Response): Promise<FetchEnd> {
-        const resourceUrl = response.url();
-        const hops = response.request()
-            .redirectChain()
-            .map((request) => {
-                return request.url();
-            });
-        const originalUrl = hops[0] || resourceUrl;
-
-        const networkRequest: Request = {
-            headers: normalizeHeaders(response.request().headers() as any) as HttpHeaders,
-            url: originalUrl
-        };
-
-        const element = await this.getElementFromResponse(response);
-        const [content, rawContent] = await Promise.all([
-            response.text(),
-            response.buffer()
-        ])
-            .catch((e) => {
-                return ['', Buffer.alloc(0)];
-            });
-
-        const body = {
-            content,
-            rawContent: rawContent || Buffer.alloc(0),
-            rawResponse: this.getRawResponse(response)
-        };
-
-        const responseHeaders = normalizeHeaders(response.headers() as any) as HttpHeaders;
-        const { charset, mediaType } = getContentTypeData(element, originalUrl, responseHeaders, body.rawContent);
-
-        const networkResponse = {
-            body,
-            charset: charset!,
-            headers: responseHeaders,
-            hops,
-            mediaType: mediaType!,
-            statusCode: response.status(),
-            url: response.url()
-        };
-
-        const data: FetchEnd = {
-            element,
-            request: networkRequest,
-            resource: resourceUrl,
-            response: networkResponse
-        };
-
-        return data;
-    }
-
-    /** Returns the HTMLElement that initiated a request */
-    private getElementFromResponse(source: puppeteer.Response | puppeteer.Request): HTMLElement | null {
-        const request = 'request' in source ?
-            source.request() :
-            source;
-
-        const redirectChain = request.redirectChain();
-        const requestUrl = redirectChain && redirectChain.length > 0 ?
-            redirectChain[0].url() :
-            source.url();
-
-        /*
-         * TODO: Check what happens with prefetch, etc.
-         * `type` can be "parser", "script", "preload", and "other": https://chromedevtools.github.io/debugger-protocol-viewer/tot/Network/#type-Initiator
-         */
-        // The doesn't seem to be an initiator in puppeteer :/
-        if (this._dom && requestUrl.startsWith('http')) {
-            return getElementByUrl(this._dom, requestUrl, this._finalHref);
-        }
-
-        return null;
-    }
-
-    /**
-     * Manually download the site's favicon:
-     *
-     * * uses the `src` attribute of `<link rel="icon">` if present.
-     * * uses `favicon.ico` and the final url after redirects.
-     */
-    private async getFavicon(dom: HTMLDocument) {
-        const element = (await dom.querySelectorAll('link[rel~="icon"]'))[0];
-        const href = (element && element.getAttribute('href')) || '/favicon.ico';
-
-        try {
-            debug(`resource ${href} to be fetched`);
-            const fullFaviconUrl = this._finalHref + href.substr(1);
-
-            await this._engine.emitAsync('fetch::start', { resource: fullFaviconUrl });
-
-            const content = await this.fetchContent(new URL(fullFaviconUrl));
-
-            const data: FetchEnd = {
-                element: null,
-                request: content.request,
-                resource: content.response.url,
-                response: content.response
-            };
-
-            await this._engine.emitAsync('fetch::end::image', data);
-        } catch (error) {
-            const event: FetchError = {
-                element,
-                error,
-                hops: [],
-                resource: href
-            };
-
-            await this._engine.emitAsync('fetch::error', event);
-        }
-    }
-
-    private getRawResponse(response: puppeteer.Response) {
-        const connector = this; // eslint-disable-line
-
-        return async function (this: any) {
-
-            const that = this; // eslint-disable-line
-
-            if (that._rawResponse) {
-                return that._rawResponse;
-            }
-
-            const rawContent = await response.buffer();
-            const responseHeaders = normalizeHeaders(response.headers())!;
-
-            if (rawContent && rawContent.length.toString() === responseHeaders['content-length']) {
-                // Response wasn't compressed so both buffers are the same
-                return rawContent;
-            }
-
-            const requestHeaders = response.request().headers();
-            const responseUrl = response.url();
-
-            /*
-             * Real browser connectors automatically request using HTTP2. This spec has
-             * [`Pseudo-Header Fields`](https://tools.ietf.org/html/rfc7540#section-8.1.2.3):
-             * `:authority`, `:method`, `:path` and `:scheme`.
-             *
-             * An example of request with those `Pseudo-Header Fields` to google.com:
-             *
-             * ```
-             * :authority:www.google.com
-             * :method:GET
-             * :path:/images/branding/googlelogo/2x/googlelogo_color_120x44dp.png
-             * :scheme:https
-             * accept:text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*\/*;q=0.8
-             * accept-encoding:gzip, deflate, br
-             * ...
-             * ```
-             *
-             * The `request` module doesn't support HTTP2 yet: https://github.com/request/request/issues/2033
-             * so the request need to be transformed to valid HTTP 1.1 ones, basically removing those headers.
-             *
-             */
-
-            const validHeaders = Object.entries(requestHeaders).reduce((final, [key, value]) => {
-                if (key.startsWith(':')) {
-                    return final;
-                }
-
-                final[key] = value;
-
-                return final;
-            }, {} as puppeteer.Headers);
-
-            return connector.fetchContent(responseUrl, validHeaders)
-                .then((result) => {
-                    const { response: { body: { rawResponse: rr } } } = result;
-
-                    return rr();
-                })
-                .then((value) => {
-                    that._rawResponse = value;
-
-                    return value;
-                });
-        };
     }
 
     private onError(error: Error) {
@@ -234,52 +55,32 @@ export default class ChromiumConnector implements IConnector {
     }
 
     private async onRequest(request: puppeteer.Request) {
-        const requestUrl = request.url();
-        const event = { resource: requestUrl };
-
-        debug(`Request started: ${requestUrl}`);
-
         if (request.isNavigationRequest()) {
             this._headers = normalizeHeaders(request.headers())!;
-
-            await this._engine.emitAsync('fetch::start::target', event);
-        } else {
-            await this._engine.emitAsync('fetch::start', event);
         }
+
+        const { name, payload } = onRequestHandler(request);
+
+        await this._engine.emitAsync(name, payload);
     }
 
     private async onRequestFailed(request: puppeteer.Request) {
         const response = request.response();
-        const resource = request.url();
 
         if (response && response.status() >= 400) {
             // Probably a 404, 503, and such, will be handled via fetch::end
             return;
         }
 
-        if (!this._dom) {
+        const event = onRequestFailedHandler(request, this._finalHref, this._dom);
+
+        if (!event) {
             this._pendingRequests.push(this.onRequestFailed.bind(this, request));
 
             return;
         }
 
-        debug(`Request failed: ${resource}`);
-
-        const element: HTMLElement | null = this.getElementFromResponse(request);
-        const eventName = 'fetch::error';
-        const hops: string[] = request.redirectChain()
-            .map((redirect) => {
-                return redirect.url();
-            });
-
-        const event: FetchError = {
-            element,
-            error: request.failure(),
-            hops,
-            resource
-        };
-
-        await this._engine.emitAsync(eventName, event);
+        await this._engine.emitAsync(event.name, event.payload);
     }
 
     private async onResponse(response: puppeteer.Response) {
@@ -289,49 +90,30 @@ export default class ChromiumConnector implements IConnector {
 
         debug(`Response received: ${resource}`);
 
-
         if (status >= 300 && status < 400) {
             // It's a redirect so it will be handled later
             return;
         }
 
-        if (!this._dom && !isTarget) {
-            // DOM isn't loaded yet, need to queue up
+        const event = (await onResponseHandler(response, this._finalHref, this.fetchContent.bind(this), this._dom))!;
+
+        if (!event) {
             this._pendingRequests.push(this.onResponse.bind(this, response));
 
             return;
         }
 
-        // TODO: Check if it needs to be queue'd up and continue if not
-
-        const fetchEndPayload = await this.createFetchEndPayload(response);
-        /*
-         * If the target has a weird value like `application/x-httpd-php`
-         * (which translates into `unknown`) or is detected as `xml`.
-         * (e.g.: because it starts with
-         * `<?xml version="1.0" encoding="utf-8"?>` even though it has
-         * `<!DOCTYPE html>` declared after),
-         * we change the suffix to `html` so hints work properly.
-         */
-        let suffix = getType(fetchEndPayload.response.mediaType);
-        const defaults = ['html', 'unknown', 'xml'];
+        const { name, payload } = event;
 
         if (isTarget) {
-
-            if (defaults.includes(suffix)) {
-                suffix = 'html';
-
-                this._html = fetchEndPayload.response.body.content;
-                this._originalDocument = createHTMLDocument(this._html);
+            if (name === 'fetch::end::html') {
+                this._html = payload.response.body.content;
+                this._originalDocument = createHTMLDocument(this._html!);
             }
-
-            this._targetNetworkData = fetchEndPayload;
+            this._targetNetworkData = payload;
         }
 
-        const eventName = `fetch::end::${suffix}` as 'fetch::end::*';
-
-        /** Event is also emitted when status code in response is not 200. */
-        await this._engine.emitAsync(eventName, fetchEndPayload);
+        await this._engine.emitAsync(name, payload);
     }
 
     private addListeners() {
@@ -378,6 +160,50 @@ export default class ChromiumConnector implements IConnector {
         }
     }
 
+    private async initiate(target: URL) {
+        if (!isRegularProtocol(target.href)) {
+            const error = new Error(`Target protocol is not valid (is ${target.protocol})`);
+
+            (error as any).type = 'InvalidTarget';
+
+            throw error;
+        }
+
+        const { browser, page } = await launch(this._options);
+
+        this._browser = browser;
+        this._page = page;
+    }
+
+    private async processTarget() {
+        if (this._html !== undefined) {
+            const event = { resource: this._finalHref };
+
+            // QUESTION: Even if the content is blank we will receive a minimum HTML with this. Are we OK with the behavior?
+            if (this._html !== '') {
+                const html = await this._page.content();
+
+                this._dom = createHTMLDocument(html, this._originalDocument);
+
+                await traverse(this._dom, this._engine, this._page.url());
+
+                if (this._options.headless) {
+                    // TODO: Check if browser downloads favicon even if there's no content
+                    await getFavicon(this._finalHref, this._dom, this.fetchContent.bind(this), this._engine);
+                }
+            }
+
+            // Process pending requests now that the dom is ready
+            while (this._pendingRequests.length > 0) {
+                const pendingRequest = this._pendingRequests.shift()!;
+
+                await pendingRequest();
+            }
+
+            await this._engine.emitAsync('can-evaluate::script', event);
+        }
+    }
+
     public async close() {
         this.removeListeners();
 
@@ -388,24 +214,10 @@ export default class ChromiumConnector implements IConnector {
         return this._page.evaluate(code);
     }
 
-    // TODO: Add options parameter
     public async collect(target: URL): Promise<any> {
-        if (!isRegularProtocol(target.href)) {
-            const error = new Error(`Target protocol is not valid (is ${target.protocol})`);
+        await this.initiate(target);
 
-            (error as any).type = 'InvalidTarget';
-
-            throw error;
-        }
-
-        const event = { resource: target.href };
-
-        const { browser, page } = await launch(this._options);
-
-        this._browser = browser;
-        this._page = page;
-
-        await this._engine.emit('scan::start', event);
+        await this._engine.emit('scan::start', { resource: target.href });
 
         // TODO: Figure out how to execute the user tasks in here and when to subscribe to events
 
@@ -417,52 +229,17 @@ export default class ChromiumConnector implements IConnector {
 
         // This is the final URL
         this._finalHref = this._page.url();
-        event.resource = this._finalHref;
 
         debug(`Navigation complete`);
 
         // Stop network events after navigation is considered complete so no more issues are reported
-        const networkEvents: EventName[] = ['request', 'requestfailed', 'response'];
+        this.removeListeners(['request', 'requestfailed', 'response']);
 
-        // Puppeteer runs in headless mode so favicon needs to be downloaded manually
-
-        // Need to make sure the dom is loaded
-
-        this.removeListeners(networkEvents);
-
-        // Await for target ready/downloaded
-
-        // await this._targetReady;
-
-        if (this._html !== undefined) {
-            // QUESTION: Even if the content is blank we will receive a minimum HTML with this. Are we OK with the behavior?
-            if (this._html !== '') {
-                const html = await this._page.content();
-
-                this._dom = createHTMLDocument(html, this._originalDocument);
-
-                await traverse(this._dom, this._engine, this._page.url());
-
-                if (this._options.headless) {
-                    // TODO: Check if browser downloads favicon even if there's no content
-                    await this.getFavicon(this._dom);
-                }
-            }
-
-            // Process pending requests now that the dom is ready
-            while (this._pendingRequests.length > 0) {
-                const pendingRequest = this._pendingRequests.shift()!;
-
-                await pendingRequest();
-            }
-
-            // TODO: Update with the final URL
-            await this._engine.emitAsync('can-evaluate::script', event);
-        }
+        await this.processTarget();
 
         // Some other timeouts or awaits here?
 
-        await this._engine.emitAsync('scan::end', event);
+        await this._engine.emitAsync('scan::end', { resource: this._finalHref });
     }
 
     public fetchContent(target: URL | string, customHeaders?: object): Promise<NetworkData> {
@@ -498,9 +275,6 @@ export default class ChromiumConnector implements IConnector {
      * Getters
      * ------------------------------------------------------------------------------
      */
-
-
-    // TODO: Are this being used somewhere?
 
     public get dom(): HTMLDocument | undefined {
         return this._dom;
